revert PhoneNumberNormalizer changes
