+++ src/Serializer/Normalizer/PhoneNumberNormalizer.php
@@ -11,64 +11,96 @@
 
 namespace Misd\PhoneNumberBundle\Serializer\Normalizer;
 
+use libphonenumber\NumberParseException;
+use libphonenumber\PhoneNumber;
+use libphonenumber\PhoneNumberFormat;
+use libphonenumber\PhoneNumberUtil;
 use Symfony\Component\Serializer\Exception\InvalidArgumentException;
 use Symfony\Component\Serializer\Exception\UnexpectedValueException;
-use Symfony\Component\Serializer\Normalizer\ArrayDenormalizer;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
-if (class_exists(ArrayDenormalizer::class) && !method_exists(ArrayDenormalizer::class, 'setSerializer')) {
-    // Symfony >= 6.0
+/**
+ * Phone number serialization for Symfony serializer.
+ */
+class PhoneNumberNormalizer implements NormalizerInterface, DenormalizerInterface
+{
     /**
-     * Phone number serialization for Symfony serializer.
+     * Region code.
+     *
+     * @var string
      */
-    class PhoneNumberNormalizer implements NormalizerInterface, DenormalizerInterface
+    private $region;
+
+    /**
+     * Display format.
+     *
+     * @var int
+     */
+    private $format;
+
+    /**
+     * Display format.
+     *
+     * @var PhoneNumberUtil
+     */
+    private $phoneNumberUtil;
+
+    /**
+     * Constructor.
+     *
+     * @param PhoneNumberUtil $phoneNumberUtil phone number utility
+     * @param string          $region          region code
+     * @param int             $format          display format
+     */
+    public function __construct(PhoneNumberUtil $phoneNumberUtil, $region = PhoneNumberUtil::UNKNOWN_REGION, $format = PhoneNumberFormat::E164)
     {
-        use PhoneNumberNormalizerTrait;
+        $this->phoneNumberUtil = $phoneNumberUtil;
+        $this->region = $region;
+        $this->format = $format;
     }
-} else {
-    // Symfony < 6.0
+
     /**
-     * Phone number serialization for Symfony serializer.
+     * {@inheritdoc}
+     *
+     * @throws InvalidArgumentException
      */
-    class PhoneNumberNormalizer implements NormalizerInterface, DenormalizerInterface
+    public function normalize($object, $format = null, array $context = [])
     {
-        use CommonPhoneNumberNormalizerTrait;
+        return $this->phoneNumberUtil->format($object, $this->format);
+    }
 
-        /**
-         * {@inheritdoc}
-         *
-         * @throws InvalidArgumentException
-         */
-        public function normalize($object, $format = null, array $context = [])
-        {
-            return $this->doNormalize($object, $format, $context);
+    /**
+     * {@inheritdoc}
+     */
+    public function supportsNormalization($data, $format = null)
+    {
+        return $data instanceof PhoneNumber;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @throws UnexpectedValueException
+     */
+    public function denormalize($data, $class, $format = null, array $context = [])
+    {
+        if (null === $data) {
+            return;
         }
 
-        /**
-         * {@inheritdoc}
-         */
-        public function supportsNormalization($data, $format = null)
-        {
-            return $this->doSupportsNormalization($data, $format);
+        try {
+            return $this->phoneNumberUtil->parse($data, $this->region);
+        } catch (NumberParseException $e) {
+            throw new UnexpectedValueException($e->getMessage(), $e->getCode(), $e);
         }
+    }
 
-        /**
-         * {@inheritdoc}
-         *
-         * @throws UnexpectedValueException
-         */
-        public function denormalize($data, $class, $format = null, array $context = [])
-        {
-            return $this->doDenormalize($data, $class, $format, $context);
-        }
-
-        /**
-         * {@inheritdoc}
-         */
-        public function supportsDenormalization($data, $type, $format = null)
-        {
-            return $this->doSupportsDenormalization($data, $type, $format);
-        }
+    /**
+     * {@inheritdoc}
+     */
+    public function supportsDenormalization($data, $type, $format = null)
+    {
+        return 'libphonenumber\PhoneNumber' === $type && \is_string($data);
     }
 }
